+++ src/api/queue_api.rs
@@ -1,5 +1,5 @@
 
 pub struct Queue {
     pub id: String,
-    pub last_event_id: i32
+    pub last_event_id: i64
 }
